university_project/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-allauth==0.63.3
django-filter==24.2
django-modeltranslation==0.19.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
pillow==10.3.0
psycopg2-binary==2.9.6
sqlparse==0.5.0
tzdata==2024.1
